refactor(frontend): split api.ts into domain modules; add 2026-05 roadmap
The 1294-line frontend/src/api.ts is broken into 14 focused modules under
frontend/src/api/ (client, auth, users, library, covers, streaming, tracks,
uploads, playlists, radio, genre, server, logs, backup) with a barrel that
preserves the public surface, so existing `from "../api"` imports keep working
unchanged. Replaces the prior REFACTORING_ROADMAP.md (mostly delivered) with a
fresh 2026-05 roadmap targeting the new hot spots.
